% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapf_ANN.R
\name{Gapfill_ann}
\alias{Gapfill_ann}
\title{Gap-fill using ANN}
\usage{
Gapfill_ann(data, Flux = "Flux", var1, var2 = NULL, var3 = NULL,
  win = 5, interval = 10, threshold = 1, hidden = 2,
  fail = "ave", ...)
}
\arguments{
\item{data}{a data frame that includes the flux (with NA indicating the missing data) and independent variables}

\item{Flux}{a string indicates the column name for the flux variable to be gap-filled}

\item{var1}{a string indicates the column name for the first variable}

\item{var2}{a string indicates the column name for the second variable, default: NULL}

\item{var3}{a string indicates the column name for the third variable, default: NULL}

\item{win}{a number indicates the required sampling window length around each gap (total number in two sides), unit: days (default: 5)}

\item{interval}{a number indicates the temporal resolution of the measurements in the dataset, unit: minutes (default: 10)}

\item{threshold}{a numeric value specifying the threshold for the partial derivatives of the error function as stopping criteria}

\item{hidden}{a vector of integers specifying the number of hidden neurons (vertices) in each layer}

\item{fail}{a string or a number, what to do when model fails to converge:
1. use the mean value in the sampling window to fill the gap ("ave", default), or
2. use any value assigned here to fill the gap (e.g., 9999, NA, etc.)}

\item{...}{other arguments pass to `neuralnet`}
}
\value{
A data frame that includes the original data, gap-filled data ("filled")
and a "mark" column that indicates the value in each row of the "filled" is either:
1. original, 2. gap-filled, or 3. failed to converge
}
\description{
This function automatically gap-fills the missing data points (marked as "NA") in the flux dataset
using artificial neural networks that take up to three variables as inputs.
}
\examples{
# read example data
df <- read.csv(file = system.file("extdata", "Soil_resp_example.csv", package = "FluxGapsR"),header = T)
df_filled <- Gapfill_ann(data = df,var1 = "Ts",var2 = "Ta",var3 = "Moist")
# visualize the gapfilled results
plot(df_filled$filled,col="red",ylim = c(0,11))
points(df_filled$Flux)
}
