% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapf_SSA.R
\name{Gapfill_ssa}
\alias{Gapfill_ssa}
\title{Gap-fill using SSA}
\usage{
Gapfill_ssa(
  data,
  Flux = "Flux",
  Date = "Date",
  Date_form = "ymd_hms",
  borders.wl = list(a = c(0, 10, Inf)),
  win = 5,
  L = NULL,
  interval = 10,
  fail = "ave",
  ...
)
}
\arguments{
\item{data}{a data frame that includes the flux (with NA indicating the missing data)}

\item{Flux}{a string indicates the column name for the flux variable to be gap-filled}

\item{Date}{a string indicates the column name for the date, which HAS to include the time information}

\item{Date_form}{a string indicates the format of the date, either "ymd_hms" (default), "mdy_hms" or "dmy_hms"}

\item{borders.wl}{list of numeric vectors indicate orders of the different periodicity bands to extract.
Units are the sampling frequency of the series
(needs one vector per step; default: list(a = c(0,10, Inf)) (see details in the package `spectral.methods`)).}

\item{win}{a number indicates the required sampling window length around each gap (total number in two sides), unit: days (default: 5)}

\item{L}{an integer indicates the window length for time series decomposition, see ?ssa.
If not provided, one third of the length of the selected time series is used.}

\item{interval}{a number indicates the temporal resolution of the measurements in the dataset, unit: minutes (default: 10)}

\item{fail}{a string or a number indicates what to do when model fails to converge:
1. use the mean value in the sampling window to fill the gap ("ave", default), or
2. use any value assigned here to fill the gap (e.g., 9999, NA, etc.)}

\item{...}{other arguments pass to `gapfillSSA`}
}
\value{
A data frame that includes the original data, gap-filled data ("filled")
and a "mark" column that indicates the value in each row of the "filled" is either:
1. original, 2. gap-filled, or 3. failed to converge
}
\description{
This function automatically gap-fills the missing data points (marked as "NA") in the flux dataset
using sigular spectrum analysis (SSA). The SSA is based on the algorithms in the package `spectral.methods` and `Rssa`.
}
\examples{
# read example data
df <- read.csv(file = system.file("extdata", "Soil_resp_example.csv", package = "FluxGapsR"),header = T)
df_filled <- Gapfill_ssa(data = df)
# visualize the gapfilled results
plot(df_filled$filled,col="red")
points(df_filled$Flux)
}
